improve: change switch to a dark mode switch
In this commit the generic Switch component is changed to a DarkModeSwitch. If the user toggles the switch it will change the <html> class and add (or remore) "dark" as class
